Create install_it.py
GIMP 2.8 plug-in to let a user install a scheme script or python plug-in by just choosing the file and letting GIMP work out where to put it in their personal GIMP folder.  Also sets execution permission on the copy of the python file so it runs under Linux.  Tested on Linux with GIMP 2.8.10, "should" work on other platforms.